+++ src/main/java/com/example/muaring/domain/social/exception/post/PostErrorCode.java
@@ -1,4 +1,4 @@
-package com.example.muaring.domain.social.exception;
+package com.example.muaring.domain.social.exception.post;
 
 import com.example.muaring.common.response.ErrorCode;
 import lombok.Getter;
